refactor(AvailableTimeService) : 기존의 AvailableTimeService 의 기능을 ReservationTimeService 로 병합
